requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.3
snowflake-connector-python==2.6.1
word2number==1.1